requirements.txt
streamlit
pandas
numpy
plotly
matplotlib
openpyxl
streamlit-aggrid
kaleido
statsmodels
scikit-learn
pyarrow 
fastparquet
seaborn
python-dotenv
fpdf2
lxml
soccerdata
cloudscraper
html5lib
